refactor: 由插件页面独立拥有草稿集合和渲染生命周期
