+++ desktop/frontend/tests/settings-studio-integration.test.js
@@ -132,62 +132,3 @@ test("current-character publication refreshes the bound controllers and unlocks 
   assert.equal(context.memoryState.rebinding, false);
   assert.equal(fields.characterEditorButton.disabled, false);
 });
-
-test("plugin refresh retains Memory editors and detaches results from the old generation", async () => {
-  const pluginSnapshot = { plugins: [{
-    installId: "memory", pluginId: "com.example.memory", sections: [{
-      sectionId: "archive", surface: "memory", values: {}, fields: [], actions: [],
-      collections: [{ collectionId: "entries", pageSize: 20, fields: [], filters: [], columns: [] }],
-    }],
-  }] };
-  const states = new Map();
-  let resolveOldQuery;
-  let queries = 0;
-  const context = settingsHandlers([
-    "applyRuntimePluginSnapshot", "pluginCollectionKey", "pluginCollectionRuntimeState", "queryPluginCollection", "clonePlain",
-  ], {
-    request: {}, pluginCollectionState: states, pluginState: {},
-    window: { clearTimeout() {}, setTimeout() { assert.fail("stale query scheduled work"); } },
-    initializePluginState() {},
-    renderPluginPage() {}, renderMemorySurface() {}, renderAboutComponents() {},
-    characterSwitching: false, memoryState: {}, pendingRuntimeCharacterId: () => null,
-    projectPluginActivity: () => ({}), memoryActivityBlocksCollection: () => false,
-    runtimePluginController: { collection() {
-      queries += 1;
-      return queries === 1 ? new Promise((resolve) => { resolveOldQuery = resolve; })
-        : Promise.resolve({ items: [{ itemId: "fresh" }], total: 1, nextCursor: null });
-    } },
-  });
-  context.applyRuntimePluginSnapshot(pluginSnapshot);
-  const oldPlugin = context.request.plugins.items[0];
-  const oldSection = oldPlugin.settings[0];
-  const oldCollection = oldSection.collections[0];
-  const oldState = context.pluginCollectionRuntimeState(oldPlugin, oldSection, oldCollection);
-  oldState.editor = { itemId: "note", values: { content: "未保存的记忆" } };
-  oldState.search = "旅行";
-  const pending = context.queryPluginCollection(oldPlugin, oldSection, oldCollection, { render: false });
-
-  context.applyRuntimePluginSnapshot(pluginSnapshot, { preserveDraft: true });
-  const newPlugin = context.request.plugins.items[0];
-  const newSection = newPlugin.settings[0];
-  const newCollection = newSection.collections[0];
-  const newState = context.pluginCollectionRuntimeState(newPlugin, newSection, newCollection);
-  assert.notEqual(newState, oldState);
-  assert.equal(JSON.stringify(newState.editor), JSON.stringify(oldState.editor));
-  assert.equal(newState.search, "旅行");
-  resolveOldQuery({ items: [{ itemId: "stale" }], total: 1, nextCursor: null });
-  await pending;
-  assert.equal(newState.items.length, 0);
-  await context.queryPluginCollection(oldPlugin, oldSection, oldCollection);
-  assert.equal(queries, 1, "old scheduled callbacks cannot issue new queries");
-  await context.queryPluginCollection(newPlugin, newSection, newCollection, { render: false });
-  assert.equal(newState.items[0].itemId, "fresh");
-  assert.equal(newState.editor.values.content, "未保存的记忆");
-
-  newState.editor = null;
-  context.applyRuntimePluginSnapshot(pluginSnapshot, { preserveDraft: true });
-  assert.equal([...states.values()][0].editor, null, "a clean collection must not acquire a phantom draft");
-
-  context.applyRuntimePluginSnapshot(pluginSnapshot);
-  assert.equal(states.size, 0, "an explicit discard still clears collection drafts");
-});
